move consumer to separate docker
