+++ app/start.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-exec python -m app.worker &
+sleep 10
 exec uvicorn app.main:api --host=0.0.0.0 --workers 2
