+++ consumer/start.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+sleep 10
+exec python -m consumer.worker
